+++ src/cmd/clone.rs
@@ -36,7 +36,7 @@ pub fn run(sess: &Session, path: &Path, args: &CloneArgs) -> Result<()> {
                                // Check current config for matches
     if sess.config.overrides.contains_key(dep) {
         match &sess.config.overrides[dep] {
-            config::Dependency::Path(p, _) => {
+            config::Dependency::Path(_, p, _) => {
                 Err(Error::new(format!(
                     "Dependency `{}` already has a path override at\n\t{}\n\tPlease check Bender.local or .bender.yml",
                     dep,
@@ -341,7 +341,7 @@ pub fn get_path_subdeps(
     }
     .iter()
     .filter_map(|(k, v)| match v {
-        config::Dependency::Path(p, _) => {
+        config::Dependency::Path(_, p, _) => {
             if p.starts_with(&old_path) {
                 Some((
                     k.clone(),