+++ src/cmd/fusesoc.rs
@@ -67,7 +67,8 @@ pub fn run_single(sess: &Session, args: &FusesocArgs) -> Result<()> {
                 sess.manifest.dependencies.keys().cloned().collect(),
                 IndexMap::new(),
                 version_string.clone(),
-                IndexMap::new(),
+                &IndexMap::new(),
+                &IndexMap::new(),
             )
             .flatten()),
         None => Err(Error::new("Error in loading sources")),