+++ src/cmd/parents.rs
@@ -136,10 +136,10 @@ pub fn get_parent_array(
             map.insert(
                 sess.manifest.package.name.clone(),
                 match sess.manifest.dependencies.get(dep).unwrap() {
-                    Dependency::Version(_, tgts) => tgts.clone(),
-                    Dependency::Path(_, tgts) => tgts.clone(),
-                    Dependency::GitRevision(_, _, tgts) => tgts.clone(),
-                    Dependency::GitVersion(_, _, tgts) => tgts.clone(),
+                    Dependency::Version(_, _, tgts) => tgts.clone(),
+                    Dependency::Path(_, _, tgts) => tgts.clone(),
+                    Dependency::GitRevision(_, _, _, tgts) => tgts.clone(),
+                    Dependency::GitVersion(_, _, _, tgts) => tgts.clone(),
                 },
             );
         } else {
@@ -177,10 +177,10 @@ pub fn get_parent_array(
                         map.insert(
                             pkg_name.to_string(),
                             match dep_manifest.dependencies.get(dep).unwrap() {
-                                Dependency::Version(_, tgts) => tgts.clone(),
-                                Dependency::Path(_, tgts) => tgts.clone(),
-                                Dependency::GitRevision(_, _, tgts) => tgts.clone(),
-                                Dependency::GitVersion(_, _, tgts) => tgts.clone(),
+                                Dependency::Version(_, _, tgts) => tgts.clone(),
+                                Dependency::Path(_, _, tgts) => tgts.clone(),
+                                Dependency::GitRevision(_, _, _, tgts) => tgts.clone(),
+                                Dependency::GitVersion(_, _, _, tgts) => tgts.clone(),
                             },
                         );
                     } else {